+++ zk-token-sdk/src/zk_token_proof_instruction.rs
@@ -25,7 +25,7 @@ pub enum ProofInstruction {
     ///
     CloseContextState,
 
-    /// Verify a close account zero-knowledge proof.
+    /// Verify a zero-balance proof.
     ///
     /// This instruction can be configured to optionally create a proof context state account.
     ///
@@ -39,9 +39,9 @@ pub enum ProofInstruction {
     ///   None
     ///
     /// Data expected by this instruction:
-    ///   `CloseAccountData`
+    ///   `ZeroBalanceProofData`
     ///
-    VerifyCloseAccount,
+    VerifyZeroBalance,
 
     /// Verify a withdraw zero-knowledge proof.
     ///
@@ -61,7 +61,7 @@ pub enum ProofInstruction {
     ///
     VerifyWithdraw,
 
-    /// Verify a withdraw withheld tokens zero-knowledge proof.
+    /// Verify a ciphertext-ciphertext equality proof.
     ///
     /// This instruction can be configured to optionally create a proof context state account.
     ///
@@ -75,9 +75,9 @@ pub enum ProofInstruction {
     ///   None
     ///
     /// Data expected by this instruction:
-    ///   `WithdrawWithheldTokensData`
+    ///   `CiphertextCiphertextEqualityProofData`
     ///
-    VerifyWithdrawWithheldTokens,
+    VerifyCiphertextCiphertextEquality,
 
     /// Verify a transfer zero-knowledge proof.
     ///
@@ -161,12 +161,12 @@ pub fn close_context_state(
     }
 }
 
-/// Create a `VerifyCloseAccount` instruction.
-pub fn verify_close_account(
+/// Create a `VerifyZeroBalance` instruction.
+pub fn verify_zero_balance(
     context_state_info: Option<ContextStateInfo>,
-    proof_data: &CloseAccountData,
+    proof_data: &ZeroBalanceProofData,
 ) -> Instruction {
-    ProofInstruction::VerifyCloseAccount.encode_verify_proof(context_state_info, proof_data)
+    ProofInstruction::VerifyZeroBalance.encode_verify_proof(context_state_info, proof_data)
 }
 
 /// Create a `VerifyWithdraw` instruction.
@@ -177,12 +177,12 @@ pub fn verify_withdraw(
     ProofInstruction::VerifyWithdraw.encode_verify_proof(context_state_info, proof_data)
 }
 
-/// Create a `VerifyWithdrawWithheldTokens` instruction.
-pub fn verify_withdraw_withheld_tokens(
+/// Create a `VerifyCiphertextCiphertextEquality` instruction.
+pub fn verify_ciphertext_ciphertext_equality(
     context_state_info: Option<ContextStateInfo>,
-    proof_data: &WithdrawWithheldTokensData,
+    proof_data: &CiphertextCiphertextEqualityProofData,
 ) -> Instruction {
-    ProofInstruction::VerifyWithdrawWithheldTokens
+    ProofInstruction::VerifyCiphertextCiphertextEquality
         .encode_verify_proof(context_state_info, proof_data)
 }
